moved nano id generator as util & smth else
