+++ middleware/main.go
@@ -1,5 +1,1 @@
 package middleware
-
-import "net/http"
-
-type Middleware = func(http.Handler) http.Handler
